requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
tensorflow==2.16.1
pandas==1.4.1
scikit-learn==1.4.2
nltk==3.6.5
matplotlib==3.9.0
tensorflow_datasets==4.9.6
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tqdm>=4.65.0
Pillow>=9.0.0
pathlib2>=2.3.7
numpy>=1.22.0